arq: tighten v3 frame header to 8 bytes
Dropped proto_ver (constant overhead), HAS_SNR bit (snr_raw==0 sufficient),
and shrunk ack_delay from 2 bytes to 1 byte (uint8, 10ms units, max 2.55s).

Wire layout (byte 0 = framer unchanged):
  [0] framer CRC6|packet_type  [1] subtype  [2] flags
  [3] session_id  [4] tx_seq  [5] rx_ack_seq
  [6] snr_raw  [7] ack_delay

DATAC4 usable payload: 46 bytes (+1 vs old v2 45 bytes).
CALL/ACCEPT keep their existing compact layout (12-byte callsign space).

+++ datalink_arq/arq_timing.c
@@ -48,7 +48,7 @@ void arq_timing_record_tx_end(arq_timing_ctx_t *ctx, int seq)
 }
 
 void arq_timing_record_ack_rx(arq_timing_ctx_t *ctx, int seq,
-                               uint16_t ack_delay_raw, int peer_snr_x10)
+                               uint8_t ack_delay_raw, int peer_snr_x10)
 {
     /* TODO Phase 3: compute RTT, emit HLOGT */
     (void)ctx;
